Refactored ARISTOTLE testing
